+++ global.json
@@ -1,6 +1,6 @@
 {
     "sdk": {
-        "version": "6.0.428",
+        "version": "8.0.404",
         "rollForward": "latestFeature"
     }
 }
