+++ src/client/Microsoft.Identity.Client/Instance/Validation/AdfsWebFingerResponse.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using Microsoft.Identity.Client.OAuth2;
 #if SUPPORTS_SYSTEM_TEXT_JSON
-using Microsoft.Identity.Client.Platforms.net8;
+using Microsoft.Identity.Client.Platforms.net;
 using System.Text.Json.Nodes;
 using JsonProperty = System.Text.Json.Serialization.JsonPropertyNameAttribute;
 #else
