+++ src/client/Microsoft.Identity.Client/ManagedIdentity/ManagedIdentityErrorResponse.cs
@@ -3,7 +3,7 @@
 
 using System;
 #if SUPPORTS_SYSTEM_TEXT_JSON
-using Microsoft.Identity.Client.Platforms.net8;
+using Microsoft.Identity.Client.Platforms.net;
 using JsonProperty = System.Text.Json.Serialization.JsonPropertyNameAttribute;
 #else
 using Microsoft.Identity.Json;
