+++ src/client/Microsoft.Identity.Client/Platforms/Android/AndroidPlatformProxy.cs
@@ -17,11 +17,7 @@ namespace Microsoft.Identity.Client.Platforms.Android
     /// <summary>
     /// Platform / OS specific logic.  No library (ADAL / MSAL) specific code should go in here.
     /// </summary>
-#if MAUI
     [Preserve(AllMembers = true)]
-#else
-    [global::Android.Runtime.Preserve(AllMembers = true)]
-#endif
     internal class AndroidPlatformProxy : AbstractPlatformProxy
     {
         internal const string AndroidDefaultRedirectUriTemplate = "msal{0}://auth";
