+++ src/client/Microsoft.Identity.Client/Platforms/Android/SystemWebview/AuthenticationActivity.cs
@@ -24,11 +24,7 @@ namespace Microsoft.Identity.Client.Platforms.Android.SystemWebview
     /// <summary>
     /// </summary>
     [Activity(Name = "microsoft.identity.client.AuthenticationActivity", Exported=true)]
-#if MAUI
     [Preserve(AllMembers = true)]
-#else
-    [global::Android.Runtime.Preserve(AllMembers = true)]
-#endif
     internal class AuthenticationActivity : Activity
     {
         internal static RequestContext RequestContext { get; set; }
