+++ src/client/Microsoft.Identity.Client/Platforms/net/JsonObjectAttribute.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.Identity.Client.Platforms.net8
+namespace Microsoft.Identity.Client.Platforms.net
 {
     /// <summary>
     /// Dummy class to mimic Microsoft.Identity.Json.JsonObjectAttribute on Net 6 platform to reduce the number of compilation flags in the code
