+++ src/client/Microsoft.Identity.Client/TokenCache.Notifications.cs
@@ -72,9 +72,7 @@ namespace Microsoft.Identity.Client
         /// delegate to deserialize a cache entry for the application and accounts specified in the <see cref="TokenCacheNotificationArgs"/>.
         /// See https://aka.ms/msal-net-token-cache-serialization
         /// </summary>
-        /// <param name="beforeAccess">Delegate set in order to handle the cache deserialization</param>
-        /// <remarks>In the case where the delegate is used to deserialize the cache, it might
-        /// want to call <see cref="Deserialize(byte[])"/></remarks>
+        /// <param name="beforeAccess">Delegate set in order to handle the cache deserialization</param>        
 #if !SUPPORTS_CUSTOM_CACHE
         [System.ComponentModel.EditorBrowsable(System.ComponentModel.EditorBrowsableState.Never)]
 #endif
@@ -89,10 +87,7 @@ namespace Microsoft.Identity.Client
         /// delegate to serialize a cache entry for the application and accounts specified in the <see cref="TokenCacheNotificationArgs"/>.
         /// See https://aka.ms/msal-net-token-cache-serialization
         /// </summary>
-        /// <param name="afterAccess">Delegate set in order to handle the cache serialization in the case where the <see cref="TokenCache.HasStateChanged"/>
-        /// member of the cache is <c>true</c></param>
-        /// <remarks>In the case where the delegate is used to serialize the cache entirely (not just a row), it might
-        /// want to call <see cref="Serialize()"/></remarks>
+        /// <param name="afterAccess">Delegate set in order to handle the cache serialization</param>
 #if !SUPPORTS_CUSTOM_CACHE
         [System.ComponentModel.EditorBrowsable(System.ComponentModel.EditorBrowsableState.Never)]
 #endif
