+++ src/client/Microsoft.Identity.Client/json/Serialization/DefaultContractResolver.cs
@@ -1406,7 +1406,7 @@ namespace Microsoft.Identity.Json.Serialization
             IValueProvider valueProvider;
 
             // UNITY flag based on https://github.com/jilleJr/Newtonsoft.Json-for-Unity
-#if !(PORTABLE40 || PORTABLE || DOTNET || ANDROID || iOS || MAC || NETSTANDARD2_0 || UNITY || UAP10_0_17763)
+#if !(PORTABLE40 || PORTABLE || DOTNET || ANDROID || iOS || NETSTANDARD2_0 || UNITY || UAP10_0_17763)
             if (DynamicCodeGeneration)
             {
                 valueProvider = new DynamicValueProvider(member);
