+++ tests/Microsoft.Identity.Test.Integration.netcore/HeadlessTests/CancellationTests.cs
@@ -8,7 +8,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client;
 using Microsoft.Identity.Client.Internal.Logger;
-using Microsoft.Identity.Client.Platforms.netcore;
+using Microsoft.Identity.Client.Platforms.netstandard;
 using Microsoft.Identity.Test.Common.Core.Helpers;
 using Microsoft.Identity.Test.Integration.Infrastructure;
 using Microsoft.Identity.Test.Unit;
